pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "athena"
version = "0.1.0"
description = "Adaptive Transformer for Hierarchical Efficient Neural Adaptation"
readme = "README.md"
requires-python = ">=3.9"
license = "Apache-2.0"
authors = [
    { name = "ATHENA Contributors" }
]
dependencies = [
    "torch>=2.3.0",
    "transformers>=4.37.0",
    "accelerate>=0.27.0",
    "bitsandbytes>=0.41.0",
    "ray[tune]>=2.9.0",
    "pyyaml>=6.0.1",
    "jsonnet>=0.20.0",
    "evaluate>=0.4.1",
    "wandb>=0.16.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
train = [
    "datasets>=2.16.0",
    "tensorboard>=2.15.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
]

[project.scripts]
athena-train = "athena.scripts.train:main"
athena-eval = "athena.scripts.evaluate:main"
athena-export = "athena.scripts.export_peft_weights:main"

[tool.hatch.build.targets.wheel]
packages = ["athena"] 